requirements.txt
flask==2.2.3
flask-cors==3.0.10
psycopg2-binary==2.9.5
boto3==1.26.84
botocore==1.29.84
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
